+++ js/config.js
@@ -1,0 +1,36 @@
+const NUMBER_OF_PHOTOS = 25;
+const MIN_LIKES = 15;
+const MAX_LIKES = 250;
+const DESCRIPTIONS = [
+  'На фото запечатлен потрясающий вид с воздуха на нашу планету из космоса.',
+  'Искривление Земли заметно, на изображении преобладают оттенки синего, зеленого и коричневого.',
+  'Видны большие водоемы, в том числе океаны, моря и реки, а также обширные земельные просторы, от пустынь до лесов и гор. Континенты узнаваемы, с четко видимыми береговыми линиями и границами, давая ощущение масштаба и разнообразия нашего мира',
+  'Солнечный свет светит на Земле, вызывая теплое свечение и создавая яркие контрасты между светом и тенью.',
+  'Фотография - напоминание о красоте и сложности нашей планеты, и призыв сохранить и защитить ее для будущих поколений.',
+  'На фото изображён панорамный вид на обширный и разнообразный мир, запечатлённый с высокой точки зрения.',
+  'С этой точки зрения, можно увидеть обширный ландшафт скатывающихся холмов, пышных лесов, извилистых рек и возвышающихся горных хребтов, простирающихся, насколько может видеть глаз.',
+  'Цвета сцены насыщены и разнообразны, с оттенками зеленого, синего, коричневого и белого, доминирующими на виде. ',
+  'Вдали можно увидеть несколько городов, каждый со своей уникальной архитектурой и городскими ландшафтами, а также суетные порты и гавани, соединяющие разные части света.',
+  'На фотографии запечатлена суть мировой красоты, сложности и разнообразия, напоминающие нам о множестве чудес, которые можно встретить на нашей планете.'
+];
+const NAMES = [
+  'Александр',
+  'Андрей',
+  'Алексей',
+  'Владимир',
+  'Вадим',
+  'Денис',
+  'Кирилл',
+  'Илья',
+  'Роман'
+];
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export {NUMBER_OF_PHOTOS,MIN_LIKES,MAX_LIKES,DESCRIPTIONS,NAMES,MESSAGES};
